Typescriptify MetadataTemplate
